refactor(server): rm/mv unused routes
This commit completely removes some of the deprecated inventory and
purchasing server routes.  In the process, `categorised/` has been
removed.  The server is now much cleaner in terms of file structure,
allowing devs to iterate on the stucture they want without having all
the old modules getting in the way.

This commit specifically removes the `donor`s and `donation`s from the
database and server architecture.
