+++ server/controllers/finance/creditors.js
@@ -1,18 +1,35 @@
+/**
+ * @module controllers/finance/creditors
+ *
+ * @description
+ * This file contains lookup routes for creditors, as needed by the complex
+ * journal vouchers page.
+ *
+ * @todo - this page is lacking integration tests
+ *
+ * @requires db
+ * @requires NotFound
+ */
+
 const db = require('../../lib/db');
 const NotFound = require('../../lib/errors/NotFound');
+
+exports.list = list;
+exports.detail = detail;
 
 /**
  * GET /creditors
  * @todo integration tests for this function
  */
-exports.list = function list(req, res, next) {
+function list(req, res, next) {
   'use strict';
 
-  var sql =
-    `SELECT BUID(c.uuid) as uuid, c.text, cg.name, BUID(c.group_uuid) as group_uuid,
+  const sql = `
+    SELECT BUID(c.uuid) as uuid, c.text, cg.name, BUID(c.group_uuid) as group_uuid,
       a.id AS account_id, a.number
     FROM creditor AS c JOIN creditor_group AS cg JOIN account AS a
-      ON c.group_uuid = cg.uuid AND cg.account_id = a.id;`;
+      ON c.group_uuid = cg.uuid AND cg.account_id = a.id;
+  `;
 
   db.exec(sql)
   .then(function (rows) {
@@ -20,27 +37,28 @@ exports.list = function list(req, res, next) {
   })
   .catch(next)
   .done();
-};
+}
 
 /**
  * GET /creditors/:uuid
  * @todo integration tests for this function
  */
-exports.detail = function detail(req, res, next) {
+function detail(req, res, next) {
   'use strict';
 
-  var sql =
-    `SELECT BUID(c.uuid) as uuid, c.text, cg.name, BUID(c.group_uuid) as group_uuid,
+  const sql = `
+    SELECT BUID(c.uuid) as uuid, c.text, cg.name, BUID(c.group_uuid) as group_uuid,
       a.id AS account_id, a.number
     FROM creditor AS c JOIN creditor_group AS cg JOIN account AS a
       ON c.group_uuid = cg.uuid AND cg.account_id = a.id
-    WHERE c.uuid = ?`;
+    WHERE c.uuid = ?;
+  `;
 
   db.exec(sql, [db.bid(req.params.uuid)])
   .then(function (rows) {
     if (!rows.length) {
       throw new NotFound(
-        `Could not find creditor with uuid ${req.params.uuid}`
+        `Could not find creditor with uuid ${req.params.uuid}.`
       );
     }
 
@@ -48,4 +66,4 @@ exports.detail = function detail(req, res, next) {
   })
   .catch(next)
   .done();
-};
+}
